+++ RELEASE-NOTES.md
@@ -11,6 +11,7 @@
 * #547 ECharts: line chart "symbol" property
 * #548 ECharts: upgrade to 6.0.0
 * #549 ECharts: Allow custom themes (experimental)
+* #551 ECharts: NPE with two series reusing column names
 
 ## Release 2.0.0-M4